requirements.txt
streamlit==1.50.0
supabase==2.4.4
requests==2.32.3
openai>=1.40.0
python-dotenv
google-api-python-client==2.126.0
google-auth==2.33.0
anyio==4.6.2.post1
mcp==1.21.0
sentence-transformers>=3.0.0,<4
transformers>=4.41,<5
torch>=2.2,<2.5           
accelerate>=0.30,<1
protobuf>=5,<6